Fix deselect button appearing when there are no options.
An incorrect type assertion ignored the possibility that Option.value is
undefined. (which can happen if the <select> element has no options)

Other solutions are of course possible (ideally the type assertions would be
made unnecessary altogether). But this fix seems quick and simple and unable
to break anything else.

Fixes #242.